Live API Integration: Modified ExchangeRateUtil.kt to use Ktor for fetching live rates from https://open.er-api.com/v6/latest/USD. It now caches these rates in memory and falls back to hardcoded defaults if the network is unavailable. 2. Automatic Refresh during Sync: Updated LedgerRepository.syncNow() to refresh currency rates whenever a sync is performed. This ensures that your financial data is balanced against up-to-date market values during the background sync cycle. 3. On-Demand Refresh for Conversion: Updated LedgerRepository.setCurrency() to fetch fresh rates immediately before converting the entire database's transactions and card balances. This guarantees the most accurate conversion possible when the user changes their primary currency in Settings. also handled sync in homepage better.

+++ composeApp/src/commonMain/kotlin/com/vellum/ledger/App.kt
@@ -322,7 +322,7 @@ fun App() {
 }
 
 @Composable
-private fun ReportItem(label: String, amount: Double, color: Color) {
+private fun ReportItem(label: String, amount: Long, color: Color) {
     val currency = com.vellum.ledger.ui.theme.LocalCurrency.current
     Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
         Text(label, color = MaterialTheme.colorScheme.onSurfaceVariant)
